manager related route done

+++ src/admin/admin.service.ts
@@ -1,10 +1,13 @@
 import * as bcrypt from "bcrypt";
 import { MailerService } from "@nestjs-modules/mailer";
-import { Injectable } from "@nestjs/common";
+import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
-import { AdminLogin, AdminProfile } from "./admin.dto";
+import { AdminLogin, AdminProfile, AdminVarifyPass } from "./admin.dto";
 import { AdminEntity } from "./admin.entity";
+import { InstructorEntity } from "src/instructor/instructor.entity";
+import { StudentEntity } from "src/student/student.entity";
+import { ManagerEntity } from "src/manager/manager.entity";
 
 @Injectable()
 export class AdminService {
@@ -13,6 +16,9 @@ export class AdminService {
 
     constructor(
         @InjectRepository(AdminEntity) private adminRepo: Repository<AdminEntity>,
+        @InjectRepository(ManagerEntity) private managerRepo: Repository<ManagerEntity>,
+        @InjectRepository(InstructorEntity) private instructorRepo: Repository<InstructorEntity>,
+        @InjectRepository(StudentEntity) private studentRepo: Repository<StudentEntity>,        
         private readonly mailerService: MailerService
     ) {}
 
@@ -46,7 +52,7 @@ export class AdminService {
             name: admin.name
         });
 
-        if(user) {
+        try {
             if(admin.password == user.password)
                 return 1;
             
@@ -60,9 +66,9 @@ export class AdminService {
                     return 0;
             }
         }
-
-        else
+        catch{
             return 0;
+        }
     }
 
 
@@ -71,12 +77,13 @@ export class AdminService {
             email: acc.email
         });
 
-        if(user) {
+        try {
             this.id = user.id;
             console.log(`ID is ${this.id}`);
 
             this.pin = Math.floor(100000 + Math.random() * 900000);
             await this.mailerService.sendMail({
+                from: "blazeaxelspy@gamil.com",
                 to: acc.email,
                 subject: `Reset password Instruction from GraspWay`,
                 text: `Let's reset your password\nG-${this.pin} is you varification code\n\n*Don't share with anyone.`,
@@ -85,13 +92,13 @@ export class AdminService {
 
             return "Varification Code sent to Admin Email."
         }
-
-        else
-            return "Email not found!"
+        catch {
+            return "Email not found!";
+        }
     }
 
 
-    async varifyPass(admin: any) {
+    async varifyPass(admin: AdminVarifyPass) {
         let isValid = false;
         if(admin.pin == this.pin)
             isValid = true;
@@ -112,9 +119,9 @@ export class AdminService {
 
     async getDashboard(): Promise<any> {
         const admin = await this.adminRepo.count({});
-        const manager = null;
-        const instructor = null;
-        const student = null;
+        const manager = await this.managerRepo.count({});
+        const instructor = await this.instructorRepo.count({});
+        const student = await this.studentRepo.count({});
         return `Admin Dashboard:\n
                 Admin: ${admin}
                 Manager: ${manager}
@@ -123,8 +130,12 @@ export class AdminService {
     }
 
 
-    editProfile(id: number, admin: AdminProfile): any {
-        return this.adminRepo.update(id, admin);
+    async editProfile(id: number, admin: AdminProfile) {
+        const salt = await bcrypt.genSalt();
+        admin.password = await bcrypt.hash(admin.password, salt);
+
+        this.adminRepo.update(id, admin);
+        return "Admin Profile Updated!";
     }
 
 
@@ -132,7 +143,7 @@ export class AdminService {
         const salt = await bcrypt.genSalt();
         admin.password = await bcrypt.hash(admin.password, salt);
 
-        this.adminRepo.update(id, {password: admin.password} );
+        console.log(this.adminRepo.update(id, {password: admin.password} ));
         return "Password reseted!";
     }
 
@@ -143,9 +154,18 @@ export class AdminService {
     }
 
     
-    deleteAdminByID(id: any): any {
-        this.adminRepo.delete(id);
-        return "Admin deleted!";
+    async deleteAdminByID(id: any): Promise<any> {
+        const user = await this.managerRepo.findOne({
+            where: {
+                id: id
+            }
+        })
+        if(user) {
+            this.adminRepo.delete(id);
+            return "Admin deleted!";
+        }
+        else 
+        throw new UnauthorizedException("ID not found");
     }
 
 // ------------------- Admin Related Routes [End] ---------------------//
@@ -153,28 +173,52 @@ export class AdminService {
 
 // ------------------- Manager Related Routes [Start] ---------------------//
 
-    addManagerByAdmin(manag: any): any {
-        throw new Error("Method not implemented.");
+    getmanagers(): any {
+        return this.managerRepo.find({
+            where: {
+                status: true
+            }
+        });
     }
 
-    searchManagerByAdmin(manag: any): any {
-        throw new Error("Method not implemented.");
+    approveManagerByAdmin(id: any): any {
+        return this.managerRepo.update(id, {status: true});
     }
 
-    editManagerProfileByAdmin(id: number, manag: any): any {
-        throw new Error("Method not implemented.");
+    async rejecteManagerByAdmin(id: any): Promise<any> {
+        const user = await this.managerRepo.findOne({
+            where: {
+                status: false,
+                id: id
+            }
+        })
+        
+        if(user)
+            return this.managerRepo.delete(id);
+        
+        else
+            throw new UnauthorizedException("invalid id");
     }
 
-    resetManagerPassByAdmin(id: number, manag: any): any {
-        throw new Error("Method not implemented.");
+    searchManagerByAdmin(id: any): any {
+        return this.managerRepo.find({ 
+            where: {id:id},
+            relations: {
+                admin: true,
+            },
+        });
     }
 
-    managerPermissionByAdmin(manag: any): any {
-        throw new Error("Method not implemented.");
-    }
+    async deleteManagerByID(id: any): Promise<any> {
+        const user = await this.managerRepo.findOne({
+            where: { id: id },
+        })
 
-    deleteManagerByID(id: any): any {
-        throw new Error("Method not implemented.");
+        if(user)
+            return this.managerRepo.delete(id);
+        
+        else
+            throw new UnauthorizedException("Manager not found");
     }
 
 // ------------------- Manager Related Routes [End] ---------------------//
@@ -182,10 +226,71 @@ export class AdminService {
 
 // ------------------- Instructor Related Routes [Start] ---------------------//
 
-    addInstructorbyAdmin(manag: any): any {
+    getinstructors(): any {
+        throw new Error("Method not implemented.");
+    }
+
+    addInstructorbyAdmin(id: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    rejectInstructorbyAdmin(id: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    searchInstructorbyAdmin(id: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    deleteInstructorbyAdmin(id: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    searchCoursebyAdmin(id: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    courseStatus(id: number, course: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    deleteCourseByAdmin(id: any): any {
         throw new Error("Method not implemented.");
     }
 
 // ------------------- Instructor Related Routes [End] ---------------------//
 
+
+// ------------------- Student Related Routes [Start] ---------------------//
+
+    getStudent(): any {
+        throw new Error("Method not implemented.");
+    }
+
+    searchStudentbyAdmin(id: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    setStudentStatus(id: number, Student: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    deleteStudentbyAdmin(id: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+// ------------------- Student Related Routes [End] ---------------------//
+
+
+// ------------------- Website Related Routes [Start] ---------------------//
+
+    addCatagory(cat: any): any {
+        throw new Error("Method not implemented.");
+    }
+
+    customizeCatagory(id: number, cat: any): any {
+        throw new Error("Method not implemented.");
+    }
+    
+// ------------------- Website Related Routes [End] ---------------------//
 }
